refactor(dev): organize `eslint.ts` and `jest.ts` into their own directories
