+++ packages/dev/lib/eslint/default.d.ts
@@ -1,3 +1,2 @@
 /** The default configuration for ESLint. */
 export declare const defaultEslintConfig: object;
-export declare function eslintConfig(config?: object): object;
