+++ packages/dev/lib/eslint/index.d.ts
@@ -1,0 +1,1 @@
+export declare function eslintConfig(config?: object): object;
