+++ packages/dev/lib/jest/default.d.ts
@@ -1,0 +1,3 @@
+import { Config } from "jest";
+/** The default configuration for Jest. */
+export declare const defaultJestConfig: Config;
