+++ packages/dev/lib/jest/index.d.ts
@@ -3,8 +3,6 @@ import { Config } from "jest";
 export type JestConfig = Config;
 /** Function type used to alter the Jest configuration. */
 export type AlterJestConfig = (config: JestConfig) => JestConfig;
-/** The default configuration for Jest. */
-export declare const defaultJestConfig: JestConfig;
 /**
  * Create a Jest configuration with optional additional configuration or a function that alters the configuration.
  * @param config Optional additional configuration or a function that alters the configuration.
