+++ packages/dev/src/eslint/index.ts
@@ -1,0 +1,10 @@
+import * as jsonfile from "jsonfile";
+import { defaultEslintConfig } from "./default";
+
+export function eslintConfig(config?: object): object {
+  jsonfile.writeFileSync("tsconfig.eslint.json", {
+    extends: "./tsconfig.json",
+    exclude: [],
+  });
+  return { ...defaultEslintConfig, ...config };
+}
